+++ src/main/java/pl/thedeem/intellij/dql/DQLIcon.java
@@ -2,6 +2,7 @@ package pl.thedeem.intellij.dql;
 
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.util.IconLoader;
+import pl.thedeem.intellij.common.IntelliJUtils;
 
 import javax.swing.*;
 
@@ -25,4 +26,6 @@ public interface DQLIcon {
     Icon DQL_ARRAY = AllIcons.Debugger.Db_array;
     Icon DQL_RECORD = AllIcons.FileTypes.Json;
     Icon INTENTION = AllIcons.Actions.Edit;
+    Icon GUTTER_EXECUTE_DQL = IntelliJUtils.scaleToBottomRight(AllIcons.Actions.Execute, DYNATRACE_LOGO, 0.5f);
+    Icon GUTTER_EXECUTE_SETTINGS = IntelliJUtils.scaleToBottomRight(AllIcons.General.GearPlain, DYNATRACE_LOGO, 0.5f);
 }