+++ src/main/java/pl/thedeem/intellij/dql/settings/DQLSettings.java
@@ -81,12 +81,4 @@ public class DQLSettings implements PersistentStateComponent<DQLSettingsState> {
     public void setDQLExecutionToolbarVisible(boolean enabled) {
         myState.showDqlExecutionToolbar = enabled;
     }
-
-    public boolean isDQLInjectionGutterIconVisible() {
-        return myState.showDqlInjectionGutterIcon;
-    }
-
-    public void setDQLInjectionGutterIconVisible(boolean enabled) {
-        myState.showDqlInjectionGutterIcon = enabled;
-    }
 }