+++ src/main/java/pl/thedeem/intellij/dql/settings/DQLSettingsConfigurable.java
@@ -33,7 +33,6 @@ public class DQLSettingsConfigurable implements Configurable {
                 || mySettingsComponent.isPerformingLiveValidationEnabled() != mySettings.isPerformingLiveValidationEnabled()
                 || mySettingsComponent.isUseDynatraceAutocompleteEnabled() != mySettings.isUseDynatraceAutocompleteEnabled()
                 || mySettingsComponent.isDQLExecutionToolbarVisible() != mySettings.isDQLExecutionToolbarVisible()
-                || mySettingsComponent.isDQLInjectionGutterIconVisible() != mySettings.isDQLInjectionGutterIconVisible()
                 || !Objects.equals(mySettingsComponent.getDefaultDynatraceTenant(), mySettings.getDefaultDynatraceTenant());
     }
 
@@ -45,7 +44,6 @@ public class DQLSettingsConfigurable implements Configurable {
         mySettings.setUseDynatraceAutocompleteEnabled(mySettingsComponent.isUseDynatraceAutocompleteEnabled());
         mySettings.setDQLExecutionToolbarVisible(mySettingsComponent.isDQLExecutionToolbarVisible());
         mySettings.setDefaultDynatraceTenant(mySettingsComponent.getDefaultDynatraceTenant());
-        mySettings.setDQLInjectionGutterIconVisible(mySettingsComponent.isDQLInjectionGutterIconVisible());
     }
 
     @Override
@@ -56,7 +54,6 @@ public class DQLSettingsConfigurable implements Configurable {
         mySettingsComponent.setUseDynatraceAutocompleteEnabled(mySettings.isUseDynatraceAutocompleteEnabled());
         mySettingsComponent.setDQLExecutionToolbarVisible(mySettings.isDQLExecutionToolbarVisible());
         mySettingsComponent.setDefaultDynatraceTenant(mySettings.getDefaultDynatraceTenant());
-        mySettingsComponent.setDQLInjectionGutterIconVisible(mySettings.isDQLInjectionGutterIconVisible());
     }
 
     @Override