+++ src/main/java/pl/thedeem/intellij/dql/settings/DQLSettingsState.java
@@ -17,8 +17,6 @@ public class DQLSettingsState {
     public String defaultTenantUrl;
     @OptionTag
     public boolean showDqlExecutionToolbar = true;
-    @OptionTag
-    public boolean showDqlInjectionGutterIcon = true;
 
     @Override
     public boolean equals(Object o) {
@@ -30,7 +28,6 @@ public class DQLSettingsState {
                 && useDynatraceAutocomplete == that.useDynatraceAutocomplete
                 && allowExperimentalFeatures == that.allowExperimentalFeatures
                 && showDqlExecutionToolbar == that.showDqlExecutionToolbar
-                && showDqlInjectionGutterIcon == that.showDqlInjectionGutterIcon
                 && Objects.equals(defaultTenantUrl, that.defaultTenantUrl);
     }
 
@@ -42,7 +39,6 @@ public class DQLSettingsState {
                 useDynatraceAutocomplete,
                 allowExperimentalFeatures,
                 showDqlExecutionToolbar,
-                showDqlInjectionGutterIcon,
                 defaultTenantUrl
         );
     }